refactor sample into a Sampler class and the evaluate script
