+++ mentat/conversation.py
@@ -292,8 +292,8 @@ class Conversation:
             max_tokens - tokens - config.token_buffer,
             loading_multiplier=0.5 * loading_multiplier,
         )
-        messages_snapshot.append(
-            ChatCompletionSystemMessageParam(role="system", content=code_message)
+        messages_snapshot.insert(
+            1, ChatCompletionSystemMessageParam(role="system", content=code_message)
         )
 
         # If we want to add agent specific messages in (this one didn't work too well):
